requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask~=2.3.2
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Keras~=2.12.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
numpy~=1.23.5
pandas~=2.0.2
protobuf==3.6.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests~=2.31.0
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow~=2.12.0
termcolor==1.1.0
urllib3~=1.26.16
Werkzeug==0.14.1

plotly~=5.15.0
mysql-connector-python~=8.0.33
beautifulsoup4~=4.12.2
colorama~=0.4.6